Fixed compilation error on Linux x64

+++ src/server/core/nxsl_classes.cpp
@@ -216,7 +216,7 @@ NXSL_Value *NXSL_DciClass::getAttr(NXSL_Object *object, const TCHAR *attr)
    }
    else if (!strcmp(attr, "lastPollTime"))
    {
-		value = new NXSL_Value(dci->getLastPollTime());
+		value = new NXSL_Value((INT64)dci->getLastPollTime());
    }
    else if (!strcmp(attr, "systemTag"))
    {
